refactor: move clearCart back to checkout component to avoid causing the mess with dependency cycles
don't update history inside it as it's going to be used at another page
