+++ src/components/checkout/clearCart.ts
@@ -1,6 +1,6 @@
-import { store } from '../constants';
-import { updateCartCount } from '../components/header/updateCartCount';
-import { storeCartProps } from '../components/products/item/controls';
+import { store } from '../../constants';
+import { updateCartCount } from '../header/updateCartCount';
+import { storeCartProps } from '../products/item/controls';
 
 export function clearCart(): void {
   const { cart } = store;
